Move html returning endpoints to proper class and rename it
