--- src/main/java/org/wildhamsters/battleships/GameStartController.java
@@ -1,16 +1,0 @@
-package org.wildhamsters.battleships;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-/**
- * @author Mariusz Bal
- */
-@Controller
-class GameStartController {
-
-    @GetMapping
-    String hello() {
-        return "register.html";
-    }
-}
